links are working plus backend - need to refactor for ProductDB
