+++ web/WEB-INF/web.xml
@@ -1,21 +1,24 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<web-app version="2.5" xmlns="http://java.sun.com/xml/ns/javaee" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://java.sun.com/xml/ns/javaee http://java.sun.com/xml/ns/javaee/web-app_2_5.xsd">
+<web-app version="3.0" xmlns="http://java.sun.com/xml/ns/javaee" 
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" 
+         xsi:schemaLocation="http://java.sun.com/xml/ns/javaee
+                             http://java.sun.com/xml/ns/javaee/web-app_3_0.xsd">
+
     <servlet>
-        <servlet-name>ProductMaintServlet</servlet-name>
-        <servlet-class>music.productMaint.ProductMaintServlet</servlet-class>
+        <servlet-name>ProductMaint</servlet-name>
+        <servlet-class>music.product.ProductMaintServlet</servlet-class>
     </servlet>
+
     <servlet-mapping>
-        <servlet-name>ProductMaintServlet</servlet-name>
-        <url-pattern>/displayProducts</url-pattern>
+        <servlet-name>ProductMaint</servlet-name>
+        <url-pattern>/ProductMaint</url-pattern>
     </servlet-mapping>
+
     <session-config>
-        <session-timeout>
-            30
-        </session-timeout>
+        <session-timeout>30</session-timeout>
     </session-config>
+    
     <welcome-file-list>
-        <welcome-file>
-            index.jsp
-        </welcome-file>
+        <welcome-file>index.jsp</welcome-file>
     </welcome-file-list>
 </web-app>
