+++ web/displayProducts.jsp
@@ -1,73 +1,38 @@
 <%@taglib prefix="c" uri="http://java.sun.com/jsp/jstl/core"%>
 <!DOCTYPE html>
 <html>
-<head>
-    <meta charset="utf-8">
-    <title>Product Maintenance</title>
-    <link rel="stylesheet" href="styles/main.css" type="text/css"/>
-</head>
-<body>
-    
-<h1>Products</h1>
+    <head>
+        <meta charset="utf-8">
+        <title>Product Maintenance</title>
+        <link rel="stylesheet" href="styles/main.css" type="text/css"/>
+    </head>
+    <body>
 
-<table>
-    <tr>
-        <th>Code</th>
-        <th>Description</th>
-        <th class="right">Price</th>
-        <th></th>
-        <th></th>
-    </tr>
-    <c:forEach var="product" items="${products}">
-        <tr>
-            <td>${product.code}</td>
-            <td>${product.description}</td>
-            <td class="right">${product.price}</td>
-            <td class="right"><input type="button" value="Edit"></td>
-            <td class="right"><input type="button" value="Delete"></td>
-        </tr>
-    </c:forEach>
-</table>
+        <h1>Products</h1>
 
-<table>
-    <tr>
-        <th>Description</th>
-        <th class="right">Price</th>
-        <th>&nbsp;</th>
-    </tr>
-    <tr>
-        <td>86 (the band) - True Life Songs and Pictures</td>
-        <td class="right">$14.95</td>
-        <td><form action="cart" method="post">
-                <input type="hidden" name="productCode" value="8601">
-                <input type="submit" value="Add To Cart">
-            </form><!--<a href="cart?productCode=8601">Add To Cart</a>--></td>
-    </tr>
-    <tr>
-        <td>Paddlefoot - The first CD</td>
-        <td class="right">$12.95</td>
-        <td><form action="cart" method="post">
-                <input type="hidden" name="productCode" value="pf01">
-                <input type="submit" value="Add To Cart">
-            </form></td>
-    </tr>
-    <tr>
-        <td>Paddlefoot - The second CD</td>
-        <td class="right">$14.95</td>
-        <td><form action="cart" method="post">
-                <input type="hidden" name="productCode" value="pf02">
-                <input type="submit" value="Add To Cart">
-            </form></td>
-    </tr>
-    <tr>
-        <td>Joe Rut - Genuine Wood Grained Finish</td>
-        <td class="right">$14.95</td>
-        <td><form action="cart" method="post">
-                <input type="hidden" name="productCode" value="jr01">
-                <input type="submit" value="Add To Cart">
-            </form></td>
-    </tr>
-</table>
-        
-</body>
+        <table>
+            <tr>
+                <th>Code</th>
+                <th>Description</th>
+                <th class="right">Price</th>
+                <th></th>
+                <th></th>
+            </tr>
+            <c:forEach var="product" items="${sessionScope.products}">
+                <tr>
+                    <td>${product.code}</td>
+                    <td>${product.description}</td>
+                    <td class="right">${product.price}</td>
+                    <td class="right">
+                        <p><a href="ProductMaint?action=addProduct&productCode=${product.code}"  >Edit</a></p>
+                    </td>
+                    <td class="right"><input type="button" value="Delete"></td>
+                </tr>
+            </c:forEach>
+        </table>
+        <form name="goToProductView" action="ProductMaint">
+            <input type="submit" value="Add Product" />
+            <input type="hidden" name="action" value="addProduct" />
+        </form>
+    </body>
 </html>
